+++ includes/class-odoo-core.php
@@ -12,7 +12,7 @@ class Odoo_Core {
     /**
      * Plugin version
      */
-    const VERSION = '1.250';
+    const VERSION = '1.251';
     
     /**
      * Constructor